requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
joblib==1.1.0
lyricsgenius==3.0.1
numpy==1.21.2
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
soupsieve==2.2.1
threadpoolctl==3.0.0
typing==3.7.4.3
urllib3==1.26.7
